documentAnaliser/config.py
from pydantic import BaseSettings
from dotenv import load_dotenv
import os

load_dotenv()

class Settings(BaseSettings):
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY", "")
    OPENAI_MODEL: str = os.getenv("OPENAI_MODEL", "gpt-4-turbo-preview")
    OUTPUT_DIR: str = os.getenv("OUTPUT_DIR", "output")

    class Config:
        env_file = ".env"

settings = Settings() 